Admin/requirements.txt
Django==2.0
requests==2.21.0
psutil==5.9.1
Pillow==9.2.0